Handle excess lines in records
In dialog records, the % expression indicates the number of lines that
should be allocated and read from the file by Race Into Space. Since
nothing prevents additional lines from being included, it would be easy
to introduce text that would never be read. This adds an option to raise
an error if trailing lines are encounted, or by default warns if
trailing lines contain text.